requirements.txt
Flask>=3.0.0
Flask-MySQLdb>=2.0.0
# Provides the MySQLdb module used by Flask-MySQLdb
mysqlclient>=2.2.0
Flask-Cors>=4.0.0
